+++ src/components/HomeCardNews.vue
@@ -95,8 +95,9 @@
                     class="mr-2"
                     style="margin-top: 0px"
                   >
-                    <span style="color: #000000 !important; font-weight: 700">NEW!</span>
-                  </v-chip
+                    <span style="color: #000000 !important; font-weight: 700"
+                      >NEW!</span
+                    > </v-chip
                   >{{ item.title }}
                 </h2>
               </div></v-card-text