+++ src/components/HomeResearchCard.vue
@@ -91,7 +91,9 @@
               class="mr-1"
               style="margin-top: -2px"
             >
-              <span style="color: #000000 !important; font-weight: 700">NEW!</span>
+              <span style="color: #000000 !important; font-weight: 700"
+                >NEW!</span
+              >
             </v-chip>
             {{ item.title }}
           </div></v-card-text